Add part2 question 2.12 solution
Generate react app with CRA, remove unneccesary files, creates needed component as requested in the exercise for the countries app. Create an application, in which one can look at data of various countries. The application gets the data from the endpoint all. The country to be shown is found by typing a search query into the search field such that if there are too many (over 10) countries that match the query, then the user is prompted to make their query more specific, if there are ten or fewer countries, but more than one, then all countries matching the query are shown. When there is only one country matching the query, then the basic data of the country, its flag and the languages spoken there, are shown.