Add part2 question 2.18 solution
Change the functionality so that if a number is added to an already existing user, the new number will replace the old number using the HTTP PUT method for updating the phone number. If the person's information is already in the phonebook, the application will confirm the action from the user.